// src/types/allocation.ts
export interface AllocationRow {
  id: string;
  dateOfConsultation: string;
  client: string;
  strategist: string;
  consultant: string;
  creditAnalyst: string;
  broker: string;
  bcCompleted: string;
  dateOfPvfReview: string;
  dateOfPropertySelection: string;
  updateOnContractSigning: string;
  updateOnContractSigningDate?: Date;
  dateOfFinanceAppointment: string;
  dateOfFinanceAppointmentDate?: Date;
}